requirements.txt
pandas
numpy
scikit-learn
joblib
nltk
matplotlib
seaborn
gensim
torch
transformers
sentence-transformers
pyyaml